Improvements to secure context work
