--- src/config/helpers/get-trust-store-certs.js
@@ -1,7 +1,0 @@
-function getTrustStoreCerts(envs) {
-  return Object.entries(envs)
-    ?.map(([key, value]) => (key.startsWith('TRUSTSTORE_') ? value : null))
-    .filter(Boolean)
-}
-
-export { getTrustStoreCerts }
