+++ src/config/index.js
@@ -1,6 +1,5 @@
 import convict from 'convict'
 import path from 'path'
-import { getTrustStoreCerts } from '~/src/config/helpers/get-trust-store-certs'
 
 const config = convict({
   env: {
@@ -74,11 +73,6 @@ const config = convict({
     format: String,
     default: '',
     env: 'CDP_HTTPS_PROXY'
-  },
-  trustStore: {
-    doc: 'CA Certificates',
-    format: Array,
-    default: getTrustStoreCerts(process.env)
   }
 })
