+++ src/helpers/secure-context/get-trust-store-certs.test.js
@@ -1,18 +1,15 @@
-import { getTrustStoreCerts } from '~/src/config/helpers/get-trust-store-certs'
+import { getTrustStoreCerts } from '~/src/helpers/secure-context/get-trust-store-certs'
 
 describe('#getTrustStoreCerts', () => {
   const mockProcessEnvWithCerts = {
-    TRUSTSTORE_CA_ONE: 'mock-cert-doris',
-    TRUSTSTORE_CA_TWO: 'mock-cert-poppy',
-    TRUSTSTORE_CA_THREE: 'mock-cert-sparkle',
+    TRUSTSTORE_CA_ONE:
+      'LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCm1vY2stY2VydC1kb3JpcwotLS0tLUVORCBDRVJUSUZJQ0FURS0tLS0tCg==',
     UNRELATED_ENV: 'not-a-cert'
   }
 
   test('Should provide expected result with "certs"', () => {
     expect(getTrustStoreCerts(mockProcessEnvWithCerts)).toEqual([
-      'mock-cert-doris',
-      'mock-cert-poppy',
-      'mock-cert-sparkle'
+      '-----BEGIN CERTIFICATE-----\nmock-cert-doris\n-----END CERTIFICATE-----'
     ])
   })
