+++ src/helpers/secure-context/index.js
@@ -1,0 +1,2 @@
+import { secureContext } from '~/src/helpers/secure-context/secure-context'
+export { secureContext }
